Add tests for joins functionality
